requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.5
pandas==2.1.3
python-dotenv==1.0.0
lxml==4.9.3
webdriver-manager==4.0.1
rich==13.7.0
openpyxl==3.1.2
playwright==1.40.0
playwright-stealth==1.0.5
APScheduler==3.10.4
flask==3.0.0
dnspython==2.6.1

# Optional: Resume parsing support
PyPDF2==3.0.1
python-docx==1.1.0
